Add `tools/utils.py` and move some common functionality there.
