+++ tools/docs.py
@@ -12,9 +12,9 @@ import sys
 
 from ptscripts import Context, command_group
 
-log = logging.getLogger(__name__)
+import tools.utils
 
-REPO_ROOT = pathlib.Path(__file__).resolve().parent.parent
+log = logging.getLogger(__name__)
 
 # Define the command group
 docs = command_group(
@@ -23,7 +23,7 @@ docs = command_group(
     description=__doc__,
     venv_config={
         "requirements_files": [
-            REPO_ROOT
+            tools.utils.REPO_ROOT
             / "requirements"
             / "static"
             / "ci"
