+++ tools/pkg.py
@@ -19,9 +19,9 @@ import tempfile
 import yaml
 from ptscripts import Context, command_group
 
-log = logging.getLogger(__name__)
+import tools.utils
 
-REPO_ROOT = pathlib.Path(__file__).resolve().parent.parent
+log = logging.getLogger(__name__)
 
 # Define the command group
 pkg = command_group(name="pkg", help="Packaging Related Commands", description=__doc__)
@@ -112,14 +112,14 @@ def set_salt_version(
     """
     Write the Salt version to 'salt/_version.txt'
     """
-    salt_version_file = REPO_ROOT / "salt" / "_version.txt"
+    salt_version_file = tools.utils.REPO_ROOT / "salt" / "_version.txt"
     if salt_version_file.exists():
         if not overwrite:
             ctx.error("The 'salt/_version.txt' file already exists")
             ctx.exit(1)
         salt_version_file.unlink()
     if salt_version is None:
-        if not REPO_ROOT.joinpath(".git").exists():
+        if not tools.utils.REPO_ROOT.joinpath(".git").exists():
             ctx.error(
                 "Apparently not running from a Salt repository checkout. "
                 "Unable to discover the Salt version."
@@ -133,7 +133,7 @@ def set_salt_version(
         ctx.info(f"Validating and normalizing the salt version {salt_version!r}...")
         with ctx.virtualenv(
             name="set-salt-version",
-            requirements_files=[REPO_ROOT / "requirements" / "base.txt"],
+            requirements_files=[tools.utils.REPO_ROOT / "requirements" / "base.txt"],
         ) as venv:
             code = f"""
             import sys
@@ -153,14 +153,14 @@ def set_salt_version(
                 ctx.exit(ctx.returncode)
             salt_version = ret.stdout.strip().decode()
 
-    if not REPO_ROOT.joinpath("salt").is_dir():
+    if not tools.utils.REPO_ROOT.joinpath("salt").is_dir():
         ctx.error(
             "The path 'salt/' is not a directory. Unable to write 'salt/_version.txt'"
         )
         ctx.exit(1)
 
     try:
-        REPO_ROOT.joinpath("salt/_version.txt").write_text(salt_version)
+        tools.utils.REPO_ROOT.joinpath("salt/_version.txt").write_text(salt_version)
     except Exception as exc:
         ctx.error(f"Unable to write 'salt/_version.txt': {exc}")
         ctx.exit(1)
@@ -211,7 +211,9 @@ def pre_archive_cleanup(ctx: Context, cleanup_path: str, pkg: bool = False):
 
     When running on Windows and macOS, some additional cleanup is also done.
     """
-    with open(str(REPO_ROOT / "pkg" / "common" / "env-cleanup-rules.yml")) as rfh:
+    with open(
+        str(tools.utils.REPO_ROOT / "pkg" / "common" / "env-cleanup-rules.yml")
+    ) as rfh:
         patterns = yaml.safe_load(rfh.read())
 
     if pkg:
@@ -317,7 +319,7 @@ def generate_hashes(ctx: Context, files: list[pathlib.Path]):
     name="source-tarball",
     venv_config={
         "requirements_files": [
-            REPO_ROOT / "requirements" / "build.txt",
+            tools.utils.REPO_ROOT / "requirements" / "build.txt",
         ]
     },
 )
@@ -342,20 +344,20 @@ def source_tarball(ctx: Context):
         "-m",
         "build",
         "--sdist",
-        str(REPO_ROOT),
+        str(tools.utils.REPO_ROOT),
         env=env,
         check=True,
     )
     # Recreate sdist to be reproducible
     recompress = Recompress(timestamp)
-    for targz in REPO_ROOT.joinpath("dist").glob("*.tar.gz"):
-        ctx.info(f"Re-compressing {targz.relative_to(REPO_ROOT)} ...")
+    for targz in tools.utils.REPO_ROOT.joinpath("dist").glob("*.tar.gz"):
+        ctx.info(f"Re-compressing {targz.relative_to(tools.utils.REPO_ROOT)} ...")
         recompress.recompress(targz)
     sha256sum = shutil.which("sha256sum")
     if sha256sum:
         packages = [
-            str(pkg.relative_to(REPO_ROOT))
-            for pkg in REPO_ROOT.joinpath("dist").iterdir()
+            str(pkg.relative_to(tools.utils.REPO_ROOT))
+            for pkg in tools.utils.REPO_ROOT.joinpath("dist").iterdir()
         ]
         ctx.run("sha256sum", *packages)
     ctx.run("python3", "-m", "twine", "check", "dist/*", check=True)
